ml/requirements.txt
# requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
# Remove problematic packages
# google-adk
# litellm==1.5.0